+++ packages/vue/src/shared/input/drag-original.ts
@@ -1,1 +1,6 @@
-export type DragOriginal = { x: number; y: number; parentId: string }
+export type DragOriginal = {
+  x: number
+  y: number
+  parentId: string
+  layoutPositioning?: 'AUTO' | 'ABSOLUTE'
+}